feat: Make testing easier by moving sandbox and helpers into spruce-skill-server
