+++ packages/spruce-skill-server/tests/SpruceTest.js
@@ -3,6 +3,7 @@ process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0'
 const globby = require('globby')
 const supertest = require('supertest')
 const faker = require('faker')
+const { generateSkillJWT } = require('./lib/jwt')
 
 // The base test model that all others will extend
 module.exports = basePath => {
@@ -40,14 +41,20 @@ module.exports = basePath => {
 
 		async setupMocks(options) {
 			try {
-				// const mocks = await globby([`${__dirname}/mocks/**/*Mock.js`])
 				const mocks = await globby([
+					`${__dirname}/mocks/**/*Mock.js`,
 					`${basePath}/server/tests/mocks/**/*Mock.js`
 				])
 				for (let i = 0; i < mocks.length; i += 1) {
-					// $FlowIgnore
 					const Mock = require(mocks[i])
 					const mock = new Mock(this.koa)
+					if (this.mocks[mock.key]) {
+						throw new Error(
+							`A mock with the key "${
+								mock.key
+							}" already exists. Please set a unique "key" in your mock.`
+						)
+					}
 					this.mocks[mock.key] = mock
 					await mock.setup(options)
 				}
@@ -58,14 +65,21 @@ module.exports = basePath => {
 
 		async before() {
 			await this.beforeBase()
+			if (this.mocks.sandbox) {
+				this.organization = this.mocks.sandbox.organization
+				const locationId = Object.keys(this.mocks.sandbox.locations)[0]
+				this.location = this.mocks.sandbox.locations[locationId]
+				this.skill = this.mocks.sandbox.skill
+			} else {
+				throw new Error(
+					'@sprucelabs/spruce-skill-server: SandboxMock has not been initialized. If this is deliberate you should override the before() method in your test'
+				)
+			}
 		}
 
 		async beforeBase(options?: Object) {
 			try {
-				// $FlowIgnore
-				// const { koa, server } = await require(`${__dirname}/../server`)
 				const { koa, server } = await require(`${basePath}/server/server`)
-				// const server = await require(`${__dirname}/../server`)
 				this.server = server
 				this.koa = koa
 				this.request = supertest(this.server)
@@ -92,6 +106,31 @@ module.exports = basePath => {
 			} catch (e) {
 				throw e
 			}
+		}
+
+		async triggerEvent({
+			eventName,
+			payload,
+			skill,
+			location,
+			organization,
+			user
+		}) {
+			const eventType = `get-views`
+			const token = generateSkillJWT({
+				skill,
+				location,
+				organization,
+				user,
+				payload,
+				eventType
+			})
+
+			const result = await this.request
+				.post('/hook.json')
+				.send({ data: token, event: eventType })
+
+			return result
 		}
 
 		createPhone() {
